move to proper directory name
