+++ src/main/scala/ca/schwitzer/scaladon/AccessToken.scala
@@ -1,4 +1,4 @@
-package ca.schwitzer.mastodon_api
+package ca.schwitzer.scaladon
 
 import akka.http.javadsl.model.headers.HttpCredentials
