+++ src/main/scala/ca/schwitzer/scaladon/AppCredentials.scala
@@ -1,4 +1,4 @@
-package ca.schwitzer.mastodon_api
+package ca.schwitzer.scaladon
 
 import play.api.libs.functional.syntax._
 import play.api.libs.json._
