+++ src/main/scala/ca/schwitzer/scaladon/models/Account.scala
@@ -1,4 +1,4 @@
-package ca.schwitzer.mastodon_api.models
+package ca.schwitzer.scaladon.models
 
 import org.joda.time.DateTime
 import play.api.libs.functional.syntax._
@@ -19,7 +19,7 @@ case class Account(id: Int,
                    statusesCount: Int)
 
 object Account {
-  import ca.schwitzer.mastodon_api.dateReads
+  import ca.schwitzer.scaladon.dateReads
 
   implicit val reads: Reads[Account] = (
     (JsPath \ "id").read[Int] and
