+++ src/main/scala/ca/schwitzer/scaladon/models/Notification.scala
@@ -1,4 +1,4 @@
-package ca.schwitzer.mastodon_api.models
+package ca.schwitzer.scaladon.models
 
 import org.joda.time.DateTime
 import play.api.libs.functional.syntax._
@@ -11,7 +11,7 @@ case class Notification(id: Int,
                         status: Option[Status])
 
 object Notification {
-  import ca.schwitzer.mastodon_api.dateReads
+  import ca.schwitzer.scaladon.dateReads
 
   implicit val reads: Reads[Notification] = (
     (__ \ "id").read[Int] and
